refactor: Import FetchError when it is used as a type
