+++ packages/cozy-stack-client/src/__tests__/CozyStackClient.spec.js
@@ -1,12 +1,13 @@
 /* eslint-env jest */
 /* global fetch */
 
-import CozyStackClient, { FetchError } from '../CozyStackClient'
+import CozyStackClient from '../CozyStackClient'
 import DocumentCollection from '../DocumentCollection'
 import JobCollection from '../JobCollection'
 import KonnectorCollection from '../KonnectorCollection'
 import jestFetchMock from 'jest-fetch-mock'
 import AppToken from '../AppToken'
+import { FetchError } from '../errors'
 
 const FAKE_RESPONSE = {
   offset: 0,
