+++ packages/cozy-stack-client/src/errors.js
@@ -7,3 +7,24 @@ export default {
   CLIENT_NOT_FOUND,
   INVALID_TOKEN
 }
+
+export class FetchError extends Error {
+  constructor(response, reason) {
+    super()
+    if (Error.captureStackTrace) {
+      Error.captureStackTrace(this, this.constructor)
+    }
+    // WARN We have to hardcode this because babel doesn't play nice when extending Error
+    this.name = 'FetchError'
+    this.response = response
+    this.url = response.url
+    this.status = response.status
+    this.reason = reason
+
+    Object.defineProperty(this, 'message', {
+      value:
+        reason.message ||
+        (typeof reason === 'string' ? reason : JSON.stringify(reason))
+    })
+  }
+}
